Squash some warnings in MSVC build.
In particular, "defined but not used" (caused by GRAPHICS_DISABLED),
double constants being truncated to floats, and implicit casts.

+++ src/ccstruct/points.h
@@ -71,7 +71,7 @@ class ICOORD
 
     ///find sq length
     float sqlength() const {
-      return xcoord * xcoord + ycoord * ycoord;
+      return (float)(xcoord * xcoord + ycoord * ycoord);
     }
 
     ///find length
@@ -95,7 +95,7 @@ class ICOORD
 
     ///find angle
     float angle() const {
-      return std::atan2(ycoord, xcoord);
+      return (float)std::atan2(ycoord, xcoord);
     }
 
     ///test equality
